backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
Pillow
paddlepaddle
paddleocr
ultralytics

streamlit
streamlit-drawable-canvas
